Implemented the IRenderer and IReadInput interfaces
Create two new classes implmenting the IRenderer and IReadInput
interfaces. The ConsoleRenderer class handles all the writing to the
console and the ConsoleReader class reads the input from the console.
Injected the ConsoleRenderer and ConsoleReader into the Engine
class through the constructor. The Scoreboard should also be injected
throught the constructor so that we implement the Strategy pattern.