Cleaning up and creating a full class diagram
Removing unused interface IInputOutputManager and the class
that implements it - ConsoleInputOutputManager. Renaming the
field factories to have a postfix of Factory. Renaming the Minesweeper
class to ConsoleMinesweeper and using it as a facade. The game Main()
is now located in the StartGame class. Removing the initial game file
that was refactored - Telerik.cs